+++ frontend/src/App.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route, Navigate } from "react-router-dom";
 //import { MapContainer, TileLayer, Marker, Popup } from "react-leaflet";
 import LoginButton from "./components/LoginButton";
 import Navbar from "./components/Navbar";
@@ -15,9 +15,17 @@ import Maps from "./pages/Maps";
 import Support from "./pages/Support";
 import Settings from "./pages/Settings";
 //import { useAuth0 } from "@auth0/auth0-react";
+let token=sessionStorage.getItem('token');
+
+// Skyddad Route-komponent
+const ProtectedRoute = ({ children }) => {
+  return token ? children : <Navigate to="/" />;
+};
+
 
 const App = () => {
   //const { isAuthenticated } = useAuth0();
+  
 
   return (
     <Router>
@@ -32,7 +40,11 @@ const App = () => {
             <Route path="/bike" element={<Bike />} />
             <Route path="/stations" element={<Stations />} />
             <Route path="/parkings" element={<Parkings />} />
-            <Route path="/users" element={<Users />} />
+            <Route path="/users" element={
+              <ProtectedRoute>
+              <Users />
+              </ProtectedRoute>
+              } />
             <Route path="/user" element={<User />} />
             <Route path="/maps" element={<Maps />} />
             <Route path="/support" element={<Support />} />
